+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog
 1.7.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- No longer test on Plone 4.1 and 4.2 and on Python 2.6.  [maurits]
+
+- Catch ``KeyError`` and ``AttributeError`` for ``getPath`` in more cases.
+  Fixes `issue #14 <https://github.com/collective/collective.catalogcleanup/issues/14>`_.
+  [maurits]
 
 
 1.7.2 (2017-09-18)